pyproject.toml
[project]
name = "rezipe"
version = "0.1.1"
description = "A way to write down your cooking/baking recipes in Python. This is an actual library, no joke (period). Made with ❤ for everyone with minimum coding literacy."
readme = "README.md"
requires-python = ">=3.12"
license = "ISC"
authors = [
    { name = "Nutchanon Ninyawee", email = "nutchanon@codustry.com" }
]
dependencies = [
    "marvin>=2.3.8",
    "pandas>=2.2.3",
    "pint-pandas>=0.7.1",
    "pint>=0.24.4",
    "pydantic>=2.10.5",
    "rich>=13.9.4",
    "tabulate>=0.9.0",
    "joblib>=1.4.2",
    "google-genai>=0.5.0",
    "google-generativeai>=0.8.3",
    "yt-dlp[curl-cffi,default]>=2025.1.15",
]

[project.urls]
repository = "https://github.com/CircleOnCircles/recipe"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
